Add in a end point for the font_paint operation.
The font_paint() operation often wants to know where the text will
end. This is vitally important in trying to determine what the next
coordinate will be when you move along the line. We can now return
the next position to use.

This change also reworks the sizing of the VDU characters so that
they work with the sizing of the Font characters.