Split up the gcontext functions and make it possible to do sizing.
The gcontext functions were all bunched together in the main gcontext
file. The VDU5 operations have been split into one file, and the
VDU shape drawing operations into another file. These are all set up
in such a way that we should be able to use Draw with them, but I cannot
find whether there ever was a library that had Draw as a back end for
it.

In any case, the string size function has been separated off to allow
it to be used as a general function for calculating the sizes of strings
for wrapping. There is now an implementation for the VDU5 back end as
well, which may make it easier to use.