Add the simple border rendering from Calendar.
The calendar gadget would have used these borders as necessary around
regions in the calendar or cells (and it's also similar to the borders
in the Mozilla front end, although I don't think the same code was
used). They have been imported raw, although they're not in the
dispatcher right now.