Split out the paint function from paintattrib. Rename to text.
The `font` routines in gcontext are now `text` prefixed. This makes
it clearer that they may not be a font.

The paint function was provided in the font implementation, but the
paintattrib function was really just a paint in the vdu5 implementation.
In theory the paintattrib function doesn't need to know about the font
itself, so could be moved to being a general function. To be decided.