Tie together the markdown parser and the slidedeck management.
It is now possible to parse the markdown file into a slidedeck,
getting all the structure we've created. Properties are not currently
supported, but they should be trivial to add in the next iteration.
It has meant splitting up some files and making it easier to work
with the slidedeck_addspan by having a 'n' variant.