Rework sliderender to be useable as a library.
The sliderender code is now able to be used as a library, callable
for individual rendering. This should allow it to be used from
elsewhere if necessary. At the moment it's just an extraction of
the test code and structures, with some extra things added, but
it's ready to be used for more structured rendering. The interface
might not be quite right with respect to the text block regions,
but that's ok - we can work on that later.