Add test CI configuration for the presentation system.
The gitlab-ci.yml will configure the way that we run the tests; at
the moment that means that we run the main makefile that builds
everything. Unfortunately we don't have the Markdown library exported,
so it won't get too far. But it will at least get some way through
the process.