Update the slide rendering to lay out the block regions better.
Block regions were being sized incorrectly - they were not sized by
the tallest span in them, but by the tallest span over the preceding
blocks. This didn't give the correct rendering when we were trying
to lay out different fonts on a line. The algorithm has changed and
it's not ideal because it now requires 3 passes through the layout
phase and one pass through the render phase. This is annoying, but
the textblock/sliderender interface would need a bit of a rework
to make it possible to do better.