Update talk and test file with recent changes.
The talk and the test file have been updated to use more features.
The talk in particular uses the larger mode 50; it's possible that
we should use a scale factor to try to make the content fit the
screen a little better. But it's not like the presentation really
needs to worry about that at the moment - if it's good enough for
me to run on the day, it meets the goal.