Add initial SlideMD support for image spans.
Image spans are now possible, allowing an image to be recorded inline
with the text spans. This means that the spans now have to have a type
and be processed slightly differently, but they are able to be recorded.
We don't currently perform any rendering for these image spans, but they
are passed through at least.