Add support for rendering background images and logos.
The background image and the logos are all handled the same way,
and have identical features for placing the images on the page in
a vaguely structured way. The intention is to make those images
able to be placed where needed without having to have any special
cases. In theory this means that the page can be watermarked, or
given a branding in the corner of the page.