Make centre logo render first; add fortify memory checking.
The centre logo generally is larger than the others so you would expect
it to appear behind, rather than overlaying the other images that you
see on the page. We move its rendering to the first entry in the array,
so that it is always at the back.

*_destroy functions have been added to clean up some of the deck and
render resources that were not being cleaned up when the objects were
destroyed. This should ensure that we get a cleaner run when we are
creating and destroying things.