Add support for rendering inline images.
Inline images are a bit of a hack. They have properties set on each
span, which must be interpreted at runtime, and - worse - we need to
load each image every time we do an operation, which means 4 times
per image in each slide. We can cache things, but it's actually not
that bad right now.