Added support for templating slide properties.
Rather than having to repeat the slide properties each time you want
to restore back to an old state, it is now possible to give a slide
a name and use those properties as a template in subsequent slides.

This makes title slides, and the like, much simpler to select.