Add a way to list an index of the slides.
To know where to jump to it is useful to have an index of what slides
are present; this isn't actually that hard, because the slidedeck
is reasonably easy to step through. At the moment the present tool
always lists the slides on its run, but this could be just a switch.