Add support for padding parameters, and add to headings.
Padding parameters are now able to be extracted in a structured way,
This is used by the quote and list blocks, and the headings. This should
make them work in a consistent way and make them easier ot extend in
the future.

There is also a separate parser for border blocks, but that has not yet
been used within the configuration.