Add some speaker notes, some pictures and firm up the main talk.
The speaker notes are just something I'll look at whilst I'm doing the
presentation - I don't have the ability to render them as well. At
least not right now.

A bunch of pictures have been added which show off some of the features
of the system so that people can see what's actually going on with the
service, and what I was doing.

The talk has been given some more obvious transition points in the
slides so that it's easier to focus on what I'm saying, rather than
having people read ahead.