Update talk, and expand the speaker notes.
The speaker notes script is now nearly complete. There's a few sections
that haven't been done, but in general it's all there. I may need to cut
it down. Or just talk fast. There's also the ability to expand the
speaker notes to a full markdown deck with one slide per physical slide,
essentially expanding the transitions that I'd condensed in the notes.

If the speaker notes are passed through the presentation system, it will
look just like another deck of slides and render just fine.