Disable debug; add Wimp_CommandWindow to make Classic happier.
Under Classic, when run from within the Desktop the initial background
colour change would be ineffective due to the CommandWindow. We now
explicitly cancel the command window to make sure that we appear
properly.

All the debug is now disabled; hopefully this will make it less noisy
if anyone uses it.