Initial work towards building the presentation tool for AArch64.
This is just some initial work to try to make the 64bit version of
the tool. It shouldn't need much in the way of changes - the things
that we'll need to build will be things like the C library functions
that are missing and maybe some better handling of some of the
module calls (which don't work yet).